+++ flutter_sound/android/src/main/java/com/dooboolab/fluttersound/FlutterSound.java
@@ -1,5 +1,4 @@
 package com.dooboolab.fluttersound;
-import com.dooboolab.ffmpeg.FlutterSoundFFmpeg;
 /*
  * Copyright 2018, 2019, 2020, 2021 Dooboolab.
  *
@@ -33,7 +32,6 @@ public class FlutterSound
 	implements FlutterPlugin,
 	           ActivityAware
 {
-    public static final boolean FULL_FLAVOR = true;
 	//static Context ctx;
 	//static Registrar reg;
 	//static Activity androidActivity;
@@ -60,7 +58,6 @@ public class FlutterSound
 
 		FlutterSoundPlayerManager.attachFlautoPlayer ( Flauto.androidContext, registrar.messenger () );
 		FlutterSoundRecorderManager.attachFlautoRecorder ( Flauto.androidContext, registrar.messenger ()  );
-        if (FULL_FLAVOR) { FlutterSoundFFmpeg.attachFFmpegPlugin( Flauto.androidContext, registrar.messenger ()  ); }
 	}
 
 
@@ -101,7 +98,6 @@ public class FlutterSound
 		Flauto.androidContext = pluginBinding.getApplicationContext ();
 		FlutterSoundPlayerManager.attachFlautoPlayer ( Flauto.androidContext, pluginBinding.getBinaryMessenger () );
 		FlutterSoundRecorderManager.attachFlautoRecorder ( Flauto.androidContext, pluginBinding.getBinaryMessenger () );
-        if (FULL_FLAVOR) { FlutterSoundFFmpeg.attachFFmpegPlugin( Flauto.androidContext, pluginBinding.getBinaryMessenger () ); }
 	}