+++ flutter_sound/ios/Classes/FlutterSoundPlayer.mm
@@ -24,8 +24,6 @@
 #import "FlutterSoundPlayer.h"
 #import <AVFoundation/AVFoundation.h>
 #import <flutter_sound_core/FlautoPlayer.h>
-#import <flutter_sound_core/FlautoTrackPlayer.h>
-#import <flutter_sound_core/FlautoTrack.h>
 #import "FlutterSoundPlayerManager.h"
          
  
@@ -124,11 +122,7 @@
 // BE CAREFUL : TrackPlayer must instance FlautoTrackPlayer !!!!!
 - (FlutterSoundPlayer*)init: (FlutterMethodCall*)call
 {
-        NSNumber* withUI = call.arguments[@"withUI"];
-        if (withUI.intValue == 0)
-                flautoPlayer = [ [FlautoPlayer alloc] init: self];
-       else
-                flautoPlayer = [ [FlautoTrackPlayer alloc] init: self];
+        flautoPlayer = [ [FlautoPlayer alloc] init: self];
         return [super init: call]; // Init Session
 }
 
@@ -328,74 +322,6 @@
         }
         [self log: DBG msg: @"IOS:<-- startPlayer"];
 }
-
-
-- (void)startPlayerFromTrack:(FlutterMethodCall*)call result: (FlutterResult)result
-{
-         [self log: DBG msg: @"IOS:--> startPlayerFromTrack"];
-         NSMutableDictionary* trackDict = (NSMutableDictionary*) call.arguments[@"track"];
-         
-         if ([trackDict[@"dataBuffer"] class] != [NSNull class])
-         {
-                trackDict[@"dataBuffer"] = [(FlutterStandardTypedData*)trackDict[@"dataBuffer"] data];
-         }
-         
-         FlautoTrack* track = [[FlautoTrack alloc] initFromDictionary: trackDict];
-         //FlutterStandardTypedData* data = (FlutterStandardTypedData*)trackDict[@"dataBuffer"];
-         //track[@"dataBuffer"] = [data data];
-         //AVAudioPlayer* toto = [[AVAudioPlayer alloc] initWithData: [zozo data] error: nil];
-
-         
-         BOOL canPause  = [call.arguments[@"canPause"] boolValue];
-         BOOL canSkipForward = [call.arguments[@"canSkipForward"] boolValue];
-         BOOL canSkipBackward = [call.arguments[@"canSkipBackward"] boolValue];
-         NSNumber* progress = (NSNumber*)call.arguments[@"progress"];
-         NSNumber* duration = (NSNumber*)call.arguments[@"duration"];
-         bool removeUIWhenStopped  = [call.arguments[@"removeUIWhenStopped"] boolValue];
-         bool defaultPauseResume  = [call.arguments[@"defaultPauseResume"] boolValue];
-         bool b = [flautoPlayer startPlayerFromTrack: track canPause: canPause canSkipForward: canSkipForward
-                       canSkipBackward: canSkipBackward progress: progress duration: duration removeUIWhenStopped: removeUIWhenStopped defaultPauseResume: defaultPauseResume];
-
-        if (b)
-        {
-   
-                        result([self getPlayerStatus]);
-        } else
-        {
-                        result([FlutterError
-                        errorWithCode:@"Audio Player"
-                        message:@"startPlayer failure"
-                        details:nil]);
-        }
-        [self log: DBG msg: @"IOS:<-- startPlayerFromTrack"];
-}
-
-- (void)nowPlaying:(FlutterMethodCall*)call result: (FlutterResult)result
-{
-                       result([FlutterError
-                                errorWithCode:@"Audio Player"
-                                message:@"Now Playing failure"
-                                details:nil]);
-                                
-                               // - (void)nowPlaying: (FlautoTrack*)track canPause: (bool)canPause canSkipForward: (bool)canSkipForward canSkipBackward: (bool)canSkipBackward
-               // defaultPauseResume: (bool)defaultPauseResume progress: (NSNumber*)progress duration: (NSNumber*)duration
-}
-
-
-- (void)setUIProgressBar:(FlutterMethodCall*)call result: (FlutterResult)result;
-{
-         NSNumber* progress = (NSNumber*)call.arguments[@"progress"];
-         NSNumber* duration = (NSNumber*)call.arguments[@"duration"];
-         double x = [ progress doubleValue];
-         progress = [NSNumber numberWithFloat: x/1000.0];
-         double y = [ duration doubleValue];
-         duration = [NSNumber numberWithFloat: y/1000.0];
-
-
-         [flautoPlayer setUIProgressBar: progress duration:duration];
-         result([self getPlayerStatus]);
-}
-