+++ doc/CHANGELOG.md
@@ -7,6 +7,12 @@ permalink: changelog.html
 summary: The Changelog of The &tau; Project.
 toc: false
 ---
+## 9.0.0
+
+- Remove the Service declaration in the Android code
+- Remove the Track Player : startPlayerFromTrack(), nowPlaying(), setUIProgressBar().
+- Remove the concept of LITE and FULL flavor : now we build just one flavor without any link with FlutterFFMPEG.
+- Remove the utilities which use FFmpeg : isFFmpegAvailable(), executeFFmpegWithArguments(), getLastFFmpegReturnCode(), getLastFFmpegCommandOutput(), ffMpegGetMediaInformation(), duration(), convertFile()
 ## 8.5.0
 
 - Added read decibel from recorder on Flutter Sound on Web. [#821](https://github.com/Canardoux/flutter_sound/pull/821). Thanks to [George Amgad](https://github.com/GeorgeAmgad) for his contribution.